requirements.txt
matplotlib
pandas
numpy
mistralai
python-dotenv
seaborn
streamlit